refactor: shared implementation of Gram operator on GPU.
