+++ src/mrinufft/operators/interfaces/cufinufft.py
@@ -23,7 +23,8 @@ from mrinufft._array_compat import (
     auto_cast,
     _array_to_cupy,
 )
-from mrinufft.operators.interfaces.utils import nvtx_mark
+from mrinufft.operators.gpu_utils import nvtx_mark
+from mrinufft.operators.toeplitz import _GramOpGpuMixin
 
 CUFINUFFT_AVAILABLE = CUPY_AVAILABLE
 try:
@@ -136,7 +137,7 @@ class RawCufinufftPlan:
         self.plans[2], self.grad_plan = self.grad_plan, self.plans[2]
 
 
-class MRICufiNUFFT(FourierOperatorBase, _ToggleGradPlanMixin):
+class MRICufiNUFFT(_GramOpGpuMixin, FourierOperatorBase, _ToggleGradPlanMixin):
     """MRI Transform operator, build around cufinufft.
 
     This operator adds density estimation and compensation (preconditioning)
@@ -417,18 +418,6 @@ class MRICufiNUFFT(FourierOperatorBase, _ToggleGradPlanMixin):
             self._d2h_stream = cp.cuda.Stream(non_blocking=True)
         return self._h2d_stream, self._d2h_stream
 
-    def _coil_slice(self, i):
-        """Return the coil ``slice`` for chunk ``i`` of the ``(B*C)//T`` loop.
-
-        ``n_trans`` (``T``) is required to divide ``n_coils`` (``C``, see
-        `__init__`), so a chunk's ``T`` coils are always a contiguous,
-        non-wrapping range within a single batch: a plain slice into
-        ``self.smaps`` is a zero-copy view (unlike fancy/advanced indexing
-        with an int array, which always allocates and copies).
-        """
-        start = (i * self.n_trans) % self.n_coils
-        return slice(start, start + self.n_trans)
-
     def _host_register(self, arr, anchor=None):
         """Page-lock ARR's memory in place (no copy), for async H2D/D2H.
 
@@ -485,28 +474,6 @@ class MRICufiNUFFT(FourierOperatorBase, _ToggleGradPlanMixin):
         if not self._host_register(contig, anchor=anchor):
             return None
         return contig.reshape(shape) if shape is not None else contig
-
-    def _accumulate_coil_combine(self, img_d, i, data_batched, smaps_batched):
-        """``img_d[b] += sum_t data_batched[t] * conj(smaps_batched[t])``.
-
-        A single custom kernel: conj, multiply and the reduction over the
-        ``T`` coils all happen in one pass over memory, with no
-        intermediate arrays and no extra kernel launches -- instead of
-        the previous conj -> multiply -> sum -> add chain. It's just
-        another kernel queued on the current stream right after
-        `_op`/`_adj_op`, so it doesn't introduce any blocking sync of
-        its own.
-
-        Since ``n_trans`` is required to divide ``n_coils`` (see
-        `__init__`), a chunk never straddles a batch boundary: the batch
-        index for every one of the ``T`` coils in chunk ``i`` is the same
-        single value ``(i*T)//n_coils``, computed on the host as a plain
-        Python int (no device array/upload needed at all).
-        """
-        T = data_batched.shape[0]
-        vol = data_batched[0].size
-        b = (i * T) // self.n_coils
-        _coil_combine_kernel(data_batched, smaps_batched, b, T, vol, img_d, size=vol)
 
     def _op_sense_device(self, data, ksp_d=None):
         T, B, C = self.n_trans, self.n_batchs, self.n_coils
@@ -1022,147 +989,6 @@ class MRICufiNUFFT(FourierOperatorBase, _ToggleGradPlanMixin):
     def _adj_op(self, coeffs_d, image_d):
         return self.raw_op.type1(coeffs_d, image_d)
 
-    @with_numpy_cupy
-    def gram_op(self, data, img_d=None, toeplitz=True):
-        """Compute the Gram operator of the NUFFT.
-
-        Parameters
-        ----------
-        data: array
-            Input data array.
-        img_d: array, optional
-            Preallocated output array.
-        toeplitz: bool, default True
-            If True, use the Toeplitz method to compute the Gram operator.
-            If False, use the direct method.
-
-        Returns
-        -------
-        NDArray
-            Array with the Gram operator applied.
-        """
-        self.check_shape(image=data)
-        if not toeplitz:
-            return self.adj_op(self.op(data))
-        if self._toeplitz_kernel is None:
-            self.compute_toeplitz_kernel()
-        if self.uses_sense and is_cuda_array(data):
-            gram_func = self._gram_op_sense_device
-        elif self.uses_sense:
-            gram_func = self._gram_op_sense_host
-        elif is_cuda_array(data):
-            gram_func = self._gram_op_calibless_device
-        else:
-            gram_func = self._gram_op_calibless_host
-        ret = gram_func(data, img_d)
-        return self._safe_squeeze(ret)
-
-    def _gram_op_sense_host(self, data, img_d):
-        T, B, C = self.n_trans, self.n_batchs, self.n_coils
-        XYZ = self.shape
-        image_dataf = np.reshape(data, (B, *XYZ))
-
-        data_batched = cp.empty((T, *XYZ), dtype=self.cpx_dtype)
-        # A chunk's T coils share a single batch (see `_coil_slice`), so
-        # only one copy of that image needs to cross PCIe; broadcasting it
-        # against the T smaps directly with `cp.multiply(..., out=...)` (a
-        # single fused kernel) is much cheaper than gathering T redundant
-        # host-side copies of the same image and multiplying separately.
-        data_single_d = cp.empty((1, *XYZ), dtype=self.cpx_dtype)
-        smaps_batched = cp.empty((T, *XYZ), dtype=self.cpx_dtype)
-        padded_array = cp.empty((T, *(s * 2 for s in XYZ)), dtype=self.cpx_dtype)
-
-        img_d = cp.zeros((B, *XYZ), dtype=self.cpx_dtype)
-        for i in range(B * C // T):
-            idx_coils = self._coil_slice(i)
-            data_single_d[0].set(image_dataf[(i * T) // C])
-
-            if not self.smaps_cached:
-                smaps_batched.set(self.smaps[idx_coils].reshape((T, *XYZ)))
-            else:
-                smaps_batched = self.smaps[idx_coils].reshape((T, *XYZ))
-            cp.multiply(data_single_d, smaps_batched, out=data_batched)
-
-            self._gram_op_raw_device(data_batched, data_batched, padded_array)
-            self._accumulate_coil_combine(img_d, i, data_batched, smaps_batched)
-        img = img_d.get().reshape((B, 1, *XYZ))
-        return img
-
-    def _gram_op_sense_device(self, data, img_d):
-        T, B, C = self.n_trans, self.n_batchs, self.n_coils
-        XYZ = self.shape
-
-        image_data = cp.asarray(data)
-        image_dataf = cp.reshape(image_data, (B, *XYZ))
-        img_d = cp.zeros((B, *XYZ), dtype=self.cpx_dtype)
-        smaps_batched = cp.empty((T, *XYZ), dtype=self.cpx_dtype)
-        data_batched = cp.empty((T, *XYZ), dtype=self.cpx_dtype)
-        padded_array = cp.empty((T, *(s * 2 for s in XYZ)), dtype=self.cpx_dtype)
-        for i in range(B * C // T):
-            idx_coils = self._coil_slice(i)
-            if not self.smaps_cached:
-                smaps_batched.set(self.smaps[idx_coils].reshape((T, *XYZ)))
-            else:
-                smaps_batched = self.smaps[idx_coils].reshape((T, *XYZ))
-            # A chunk's T coils share a single batch (see `_coil_slice`):
-            # a single fused broadcast-multiply kernel straight from the
-            # caller's image view, instead of gathering (and copying) a
-            # redundant T-way repeat of it via fancy indexing.
-            cp.multiply(image_dataf[(i * T) // C], smaps_batched, out=data_batched)
-            self._gram_op_raw_device(data_batched, data_batched, padded_array)
-
-            self._accumulate_coil_combine(img_d, i, data_batched, smaps_batched)
-        img_d = img_d.reshape((B, 1, *XYZ))
-        return img_d
-
-    def _gram_op_calibless_host(self, data, img_d):
-        T, B, C = self.n_trans, self.n_batchs, self.n_coils
-        XYZ = self.shape
-        image_dataf = np.reshape(data, (B * C, *XYZ))
-        if img_d is None:
-            img_d = np.empty((B * C, *XYZ), dtype=self.cpx_dtype)
-        else:
-            img_d = img_d.reshape((B * C, *XYZ))
-        data_batched = cp.empty((T, *XYZ), dtype=self.cpx_dtype)
-        padded_array = cp.empty((T, *(s * 2 for s in XYZ)), dtype=self.cpx_dtype)
-        for i in range(B * C // T):
-            data_batched.set(image_dataf[i * T : (i + 1) * T])
-            self._gram_op_raw_device(data_batched, data_batched, padded_array)
-            img_d[i * T : (i + 1) * T] = data_batched.get()
-        img_d = img_d.reshape((B, C, *XYZ))
-        return img_d
-
-    def _gram_op_calibless_device(self, data, img_d):
-        T, B, C = self.n_trans, self.n_batchs, self.n_coils
-        XYZ = self.shape
-
-        image_data = cp.asarray(data).reshape(B * C, *XYZ)
-        padded_array = cp.empty((T, *(s * 2 for s in XYZ)), dtype=self.cpx_dtype)
-
-        if img_d is None:
-            img_d = cp.empty((B * C, *XYZ), dtype=self.cpx_dtype)
-        else:
-            img_d = img_d.reshape((B * C, *XYZ))
-
-        for i in range(B * C // T):
-            self._gram_op_raw_device(
-                image_data[i * T : (i + 1) * T],
-                img_d[i * T : (i + 1) * T],
-                padded_array,
-            )
-        img_d = img_d.reshape((B, C, *XYZ))
-        return img_d
-
-    def _gram_op_raw_device(self, in_d, out_d, padded_array=None):
-        """Apply the toeplitz Gram operator on device on a single image."""
-        # TODO Add support for batching with n_trans.
-        from mrinufft.operators.toeplitz import apply_toeplitz_kernel
-
-        cp.copyto(
-            out_d, apply_toeplitz_kernel(in_d, self._toeplitz_kernel, padded_array)
-        )
-        return out_d
-
     def data_consistency(self, image_data, obs_data):
         """Compute the data consistency estimation directly on gpu.
 
@@ -1229,7 +1055,6 @@ class MRICufiNUFFT(FourierOperatorBase, _ToggleGradPlanMixin):
         obs_batched = cp.empty((T, K), dtype=self.cpx_dtype)
 
         grad_d = cp.zeros((B, *XYZ), dtype=self.cpx_dtype)
-        grad = np.empty((B, *XYZ), dtype=self.cpx_dtype)
         inv_norm = 1.0 / self.norm_factor
         for i in range(B * C // T):
             idx_coils = self._coil_slice(i)
